:sparkles: AsyncIterDataPipe for running concurrent tasks (#5)
* :sparkles: AsyncIterDataPipe for running concurrent tasks

An asynchronous iterable-style DataPipe for processing tasks concurrently! Composition over inheritance. Subclassing from collections.abc.AsyncIterable in Python's standard library. Added some basic API docstring, and have setup some extlinks and intersphinx mappings in the docs/_config.yml file for linking to terms in the Python glossary.

* :white_check_mark: Unit test for __repr__ method of AsyncIterDataPipe

Minimal unit test to ensure that an AsyncIterDataPipe's qualified name is returned when calling repr() on it.

* :memo: The big why of using bambooflow

On where bambooflow fits in the world in relation to other libraries.