:sparkles: MapperAsyncIterDataPipe for applying custom async functions (#9)
* :sparkles: MapperAsyncIterDataPipe for applying custom async functions

An asynchronous iterable-style DataPipe for applying a custom asynchronous function over an asynchronous iterable! Uses asyncio.TaskGroup from Python 3.11+ to run several tasks concurrently. Included a doctest, added a new section in the API docs under 'Mapping DataPipes', and set show_toc_level to 3 to make it show in the right sidebar.

* :white_check_mark: Add unit test for MapperAsyncIterDataPipe

Ensure that tasks are processed concurrently, included a timer to double check that all 3 tasks (made up of 2 sub-tasks) complete in 0.5 seconds instead of 1.5 seconds!

* :goal_net: Use try-except* to catch task errors in ExceptionGroup

To better handle errors from tasks in a TaskGroup, wrap the TaskGroup context manager in a try-except* clause following PEP 654 Exception Groups. Based on the nice examples from https://github.com/jrfk/talk/tree/main/EuroPython2023. Added a unit test to ensure that a ValueError raised in 1 out of 3 tasks can be nicely captured and raised to attention.

* :memo: Document the raising of ExceptionGroup when a task errors out

Mention PEP0654 so that people know what an ExceptionGroup is, and how it could be handled.